Switch to Sphinx/reStructuredText-based docs
The current documentation building setup is fragile, hard to maintain
(we even have a custom docker image just for building the docs!), and at
the same time not very flexible.

The main reason is that various different incompatible MarkDown flavors
exist, which require incompatible processing software. reStructuredText
does not suffer from this problem: it's well-defined and extensible.
Together with Sphinx various types of documentation can be integrated,
such as API docs. In combination with ReadTheDocs, documentation can be
easily published.

This commit is huge, as I tried to remove and add the infrastructure in
one commit to keep as much history as possible (e.g. renames).