EVENT packets don't have a spec'ed format
In previous versions, the EVENT packets had their length encoded in the
packet length itself, as well as in the payload. Remove this duplication
of information. The packet length is now the only source for this
information.

Now that the special format requirements for EVENT packets are removed,
the PLAIN packet type isn't needed any more. (MAM already switched to
using EVENT packets.)