CDM: Simplify and generalize specification
The CDM module is now less or1k-specific. This is achieved by
restricting the spec to be a access module to the core's SPRs, and
leaving all other semantics up to the core spec.

Despite these efforts, the CDM is currently only targeting the or1k
core, mostly due to the lack of insight into other cores, and the
missing need. In future iterations we hope to further generalize the
spec in this direction.