Make whole documentation available as PDF in RTD
Read The Docs shows only one PDF file, we currently produce multiple
ones (which we'd like to keep). Generate one more "master" document
which contains all other content.

The only "problem" is that the top-level section is always a chapter,
i.e. what is a section in the master document is a chapter in the
individual documents.